chore: reorder changes and introduce UITestingParameters enum
